- Worked on SearchActivity.kt - Removed SearchFragment.kt - Added Vector Assets - Cleaned Up the Code at places
