+++ app/src/main/res/drawable-xhdpi/ic_music.xml
@@ -1,0 +1,18 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#ef8229" android:pathData="m487.52,492.133c-4.238,0 -7.676,-3.438 -7.676,-7.676v-130.793c0,-4.238 3.438,-7.676 7.676,-7.676s7.672,3.438 7.672,7.676v130.793c0,4.238 -3.434,7.676 -7.672,7.676zM487.52,492.133"/>
+    <path android:fillColor="#ef8229" android:pathData="m350.16,437.309c-4.238,0 -7.672,-3.438 -7.672,-7.672v-109.004c0,-4.238 3.434,-7.672 7.672,-7.672 4.238,0 7.676,3.434 7.676,7.672v109.004c0,4.234 -3.438,7.672 -7.676,7.672zM350.16,437.309"/>
+    <path android:fillColor="#ff9c43" android:pathData="m342.488,276.082v44.551l152.703,36.613v-45.594c0,-1.715 -1.172,-3.211 -2.836,-3.617l-145.262,-35.57c-2.344,-0.574 -4.605,1.203 -4.605,3.617zM342.488,276.082"/>
+    <path android:fillColor="#ef8229" android:pathData="m492.355,308.035 l-50.492,-12.363v48.789l53.328,12.785v-45.594c0,-1.715 -1.172,-3.211 -2.836,-3.617zM492.355,308.035"/>
+    <path android:fillColor="#ff9c43" android:pathData="m357.836,429.633c0,-11.328 -13.727,-20.516 -30.66,-20.516 -16.934,0 -30.66,9.188 -30.66,20.516 0,11.332 13.727,20.516 30.66,20.516 16.934,0 30.66,-9.184 30.66,-20.516zM357.836,429.633"/>
+    <path android:fillColor="#ef8229" android:pathData="m495.191,484.457c0,-11.332 -13.727,-20.516 -30.66,-20.516 -16.93,0 -30.656,9.184 -30.656,20.516s13.727,20.516 30.656,20.516c16.934,0 30.66,-9.184 30.66,-20.516zM495.191,484.457"/>
+    <path android:fillColor="#aa4b21" android:pathData="m168.348,512c-1.965,0 -3.926,-0.75 -5.426,-2.246 -2.996,-3 -2.996,-7.859 0,-10.855l14.434,-14.434c3.781,-3.781 5.863,-8.805 5.863,-14.148 0,-5.348 -2.082,-10.371 -5.863,-14.148 -3.777,-3.781 -8.801,-5.859 -14.145,-5.859 -5.348,0 -10.371,2.078 -14.148,5.859l-36.555,36.555c-22.047,22.047 -57.922,22.047 -79.973,0 -22.047,-22.047 -22.047,-57.922 0,-79.969l15.148,-15.148c2.996,-2.996 7.855,-2.996 10.852,0 3,2.996 3,7.855 0,10.852l-15.145,15.148c-7.781,7.781 -12.07,18.129 -12.07,29.133 0,11.004 4.285,21.352 12.07,29.133 7.781,7.781 18.125,12.066 29.129,12.066 11.008,0 21.352,-4.285 29.133,-12.066l36.559,-36.559c6.676,-6.676 15.555,-10.355 25,-10.355 9.441,0 18.32,3.68 25,10.355 6.676,6.68 10.355,15.559 10.355,25.004 0,9.441 -3.68,18.32 -10.355,25l-14.438,14.438c-1.496,1.496 -3.461,2.246 -5.426,2.246zM168.348,512"/>
+    <path android:fillColor="#454851" android:pathData="m54.305,413.273 l-11.434,-11.434c-7.168,-7.168 -7.168,-18.781 0,-25.945l69.441,-69.445 37.383,37.379 -69.445,69.445c-7.164,7.164 -18.781,7.164 -25.945,0zM54.305,413.273"/>
+    <path android:fillColor="#f43b45" android:pathData="m456.141,91.813c0,50.707 -41.105,91.816 -91.813,91.816s-91.813,-41.109 -91.813,-91.816 41.105,-91.813 91.813,-91.813 91.813,41.105 91.813,91.813zM456.141,91.813"/>
+    <path android:fillColor="#d82b2b" android:pathData="m295.16,152.191c1.359,1.551 2.77,3.07 4.246,4.547 35.855,35.855 93.988,35.855 129.844,0s35.855,-93.992 0,-129.848c-6.551,-6.551 -13.848,-11.898 -21.613,-16.055 18.594,34.742 13.246,78.926 -16.059,108.227 -25.449,25.449 -63.965,41.516 -96.418,33.129zM295.16,152.191"/>
+    <path android:fillColor="#d82b2b" android:pathData="m317.891,12.59c-4.625,2.715 -9.055,5.891 -13.246,9.488 7.453,23.113 26.469,51.301 52.297,77.125 25.824,25.824 54.008,44.84 77.121,52.297 3.598,-4.191 6.773,-8.621 9.488,-13.246 -21.148,-5.168 -49.984,-24.133 -75.758,-49.902 -25.773,-25.777 -44.734,-54.609 -49.902,-75.762zM317.891,12.59"/>
+    <path android:fillColor="#345863" android:pathData="m89.539,403.828 l-37.223,-37.223c-4.348,-4.348 -4.563,-11.328 -0.488,-15.934l221.105,-250.035c17.113,36.27 46.305,65.457 82.57,82.57l-250.031,221.109c-4.609,4.07 -11.586,3.859 -15.934,-0.488zM89.539,403.828"/>
+    <path android:fillColor="#1e4751" android:pathData="m355.504,183.207c-15.816,-7.465 -30.289,-17.227 -42.996,-28.867l-250.734,221.723 27.766,27.766c4.348,4.348 11.324,4.559 15.934,0.488zM355.504,183.207"/>
+    <path android:fillColor="#ffe471" android:pathData="m77.117,322.074 l56.949,56.953 27.875,-24.648 -60.176,-60.18zM77.117,322.074"/>
+    <path android:fillColor="#f2ca20" android:pathData="m98.566,343.523 l35.5,35.504 27.875,-24.648 -35.5,-35.5zM98.566,343.523"/>
+</vector>
