+++ app/src/main/res/drawable-xhdpi/ic_musician.xml
@@ -1,0 +1,31 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#9b8268" android:pathData="m184,408h64v32h-64z"/>
+    <path android:fillColor="#a98e72" android:pathData="m160,368h32v120h-32z"/>
+    <path android:fillColor="#a98e72" android:pathData="m240,368h32v120h-32z"/>
+    <path android:fillColor="#937c65" android:pathData="m160,376h31.535v10.811h-31.535z"/>
+    <path android:fillColor="#937c65" android:pathData="m240,376h31.535v10.811h-31.535z"/>
+    <path android:fillColor="#905638" android:pathData="m264,424 l8,24 24,16h15.549a15.792,15.792 0,0 0,13.139 -7.032l1.389,-2.084a15.791,15.791 0,0 0,-6.077 -22.884l-16,-16 -7.538,-8.377 -24.462,6.777z"/>
+    <path android:fillColor="#905638" android:pathData="m272,376h-112a16,16 0,0 1,-16 -16,16 16,0 0,1 16,-16h32v-8l80,-0.8v8.8a16,16 0,0 1,16 16,16 16,0 0,1 -16,16z"/>
+    <path android:fillColor="#fad7be" android:pathData="m344,256 l-40,-32 -32.845,-65.689a25.887,25.887 0,0 0,-23.155 -14.311l-8,40h8l24,56 64,48h24a16,16 0,0 0,16 -16v-16l-16,-8z"/>
+    <path android:fillColor="#c48412" android:pathData="m304,232h112v32h-112z"/>
+    <path android:fillColor="#83bd93" android:pathData="m248,296 l-72,8 -24,-128 4.97,-19.881a16,16 0,0 1,15.522 -12.119h75.508z"/>
+    <path android:fillColor="#fad7be" android:pathData="M208,24L208,24A32,32 0,0 1,240 56L240,80A32,32 0,0 1,208 112L208,112A32,32 0,0 1,176 80L176,56A32,32 0,0 1,208 24z"/>
+    <path android:fillColor="#dc9315" android:pathData="m272,296a43.266,43.266 0,0 0,-48 0h-16a48,48 0,0 1,-48 -48,48 48,0 0,1 48,-48h16a43.266,43.266 0,0 0,48 0,48 48,0 0,1 48,48 48,48 0,0 1,-48 48z"/>
+    <path android:fillColor="#dc9315" android:pathData="m472,272h-48l-16,-8v-32l16,-8h48z"/>
+    <path android:fillColor="#905638" android:pathData="m344,400 l12.406,-4.135a16,16 0,0 1,18.373 6.3l9.221,13.835 -44.422,30.8a8,8 0,0 1,-11.509 -2.613l-16.069,-28.187 7.482,-11.741z"/>
+    <path android:fillColor="#196b8e" android:pathData="m200,296v48h48l16,80 40,-8 -25.665,-96.245a32,32 0,0 0,-30.92 -23.755z"/>
+    <path android:fillColor="#905638" android:pathData="m208,24a32,32 0,0 0,-32 32v5.926a263.942,263.942 0,0 0,64 0v-5.926a32,32 0,0 0,-32 -32z"/>
+    <path android:fillColor="#69b9e7" android:pathData="m392,128h-24v-8h24a8,8 0,0 0,0 -16h-32a8,8 0,0 0,-8 8v32h-8a24,24 0,1 0,24 24v-24h24a8,8 0,0 0,0 -16z"/>
+    <path android:fillColor="#ff9201" android:pathData="m488,16h-32a8,8 0,0 0,-8 8v32h-8a24,24 0,1 0,24 24v-48h24a8,8 0,0 0,0 -16z"/>
+    <path android:fillColor="#ec6a62" android:pathData="m488,312h-32a8,8 0,0 0,-8 8v32h-8a24,24 0,1 0,24 24v-48h24a8,8 0,0 0,0 -16z"/>
+    <path android:fillColor="#ffb1b4" android:pathData="m96,256h-24v-8h24a8,8 0,0 0,0 -16h-32a8,8 0,0 0,-8 8v32h-8a24,24 0,1 0,24 24v-24h24a8,8 0,0 0,0 -16z"/>
+    <path android:fillColor="#d08b14" android:pathData="m218.835,226.357 l-39.385,-16.878a47.889,47.889 0,0 0,-18.811 31.052,31.572 31.572,0 0,0 3.758,1.795l42.065,15.3a16.819,16.819 0,0 0,12.373 -31.266z"/>
+    <path android:fillColor="#b47a11" android:pathData="m336,256s0.11,3.417 0.259,8h31.741v-8l-16,-8z"/>
+    <path android:fillColor="#d08b14" android:pathData="m283.3,294.606 l-3.565,-7.129a32,32 0,0 0,-45.581 -12.825l-23.13,10.954c-4.117,1.454 -13.022,4.9 -19.492,7.434a47.793,47.793 0,0 0,16.468 2.96h16a43.266,43.266 0,0 1,48 0,47.972 47.972,0 0,0 11.3,-1.394z"/>
+    <path android:fillColor="#37869f" android:pathData="m192,296a24,24 0,0 0,-24 24,24 24,0 0,0 24,24h9.467a47.994,47.994 0,0 0,26.625 -8.062l35.908,-23.938 48,104 32,-16 -56.261,-112.523a32,32 0,0 0,-45.582 -12.825l-23.13,10.954c-6.854,2.42 -27.027,10.394 -27.027,10.394z"/>
+    <path android:fillColor="#c48412" android:pathData="M252.001,246.562m-17.983,0a17.983,17.983 0,1 1,35.966 0a17.983,17.983 0,1 1,-35.966 0"/>
+    <path android:fillColor="#75462e" android:pathData="m208,71.848a272.223,272.223 0,0 1,-65.941 -8.086,8 8,0 1,1 3.882,-15.524 256.791,256.791 0,0 0,124.118 0,8 8,0 0,1 3.882,15.524 272.223,272.223 0,0 1,-65.941 8.086z"/>
+    <path android:fillColor="#fad7be" android:pathData="m206.462,249.623 l-42.065,-15.3a32,32 0,0 1,-19.714 -39.268l12.012,-40.04a15.459,15.459 0,0 1,14.805 -11.015,15.461 15.461,0 0,1 14.866,19.708l-10.366,36.292 42.835,18.358a16.819,16.819 0,0 1,9.691 19.538,16.819 16.819,0 0,1 -22.064,11.727z"/>
+    <path android:fillColor="#fad7be" android:pathData="m345.046,288h14.954a16,16 0,0 0,16 -16v-16l-16,-8 -16,8s1.046,32.383 1.046,32z"/>
+</vector>
