+++ app/src/main/res/drawable-xhdpi/ic_record.xml
@@ -1,0 +1,10 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#596c76" android:pathData="m244.009,0c-129.071,0 -233.704,104.633 -233.704,233.705s104.633,233.705 233.705,233.705 233.705,-104.633 233.705,-233.705 -104.634,-233.705 -233.706,-233.705zM244.009,288.195c-30.046,0 -54.489,-24.444 -54.489,-54.49s24.444,-54.49 54.489,-54.49c30.046,0 54.49,24.444 54.49,54.49s-24.444,54.49 -54.49,54.49z"/>
+    <path android:fillColor="#384949" android:pathData="m244.009,117.551c-64.047,0 -116.153,52.106 -116.153,116.154s52.106,116.154 116.153,116.154c30.858,0 58.94,-12.098 79.762,-31.797v-43.185c0,-3.168 2.161,-5.927 5.236,-6.686l27.829,-6.87c2.168,-8.855 3.326,-18.102 3.326,-27.615 0.001,-64.049 -52.105,-116.155 -116.153,-116.155zM244.009,288.195c-30.046,0 -54.489,-24.444 -54.489,-54.49s24.444,-54.49 54.489,-54.49c30.046,0 54.49,24.444 54.49,54.49s-24.444,54.49 -54.49,54.49z"/>
+    <path android:fillColor="#465a61" android:pathData="m244.009,0c-6.685,0 -13.3,0.297 -19.842,0.847 119.782,10.073 213.862,110.471 213.862,232.858s-94.08,222.785 -213.862,232.858c6.543,0.55 13.158,0.847 19.842,0.847 129.071,0 233.705,-104.633 233.705,-233.705s-104.633,-233.705 -233.705,-233.705z"/>
+    <path android:fillColor="#fe646f" android:pathData="M410.693,457.507a32.103,49.1 64.973,1 0,88.98 -41.544a32.103,49.1 64.973,1 0,-88.98 41.544z"/>
+    <path android:fillColor="#fe646f" android:pathData="m501.695,234.353c0,-4.472 -4.196,-7.758 -8.538,-6.686l-164.15,40.524c-3.075,0.759 -5.236,3.518 -5.236,6.686v67.813h0.002v97.959c-9.771,-0.691 -21.031,1.388 -32.012,6.508 -24.579,11.462 -38.43,33.781 -30.935,49.852 7.494,16.071 33.495,19.808 58.074,8.346 20.096,-9.371 33.012,-25.998 32.963,-40.536h0.008c0,-2.804 0,-81.598 0,-129.066l149.825,-36.987v-64.413z"/>
+    <path android:fillColor="#fd4755" android:pathData="m471.714,401.339h0.002z"/>
+    <path android:fillColor="#fd4755" android:pathData="m493.158,227.666 l-19.559,4.828v33.352,134.854c-0.003,0 -0.006,-0.001 -0.008,-0.001v23.023h-0.044c0.755,14.799 -12.304,32.057 -32.927,41.673 -4.492,2.095 -9.03,3.676 -13.495,4.773 11.481,3.499 26.786,2.131 41.592,-4.773 20.193,-9.416 33.133,-26.159 32.957,-40.747h0.022c0,-9.335 0,-182.783 0,-190.295 -0.001,-4.473 -4.196,-7.758 -8.538,-6.687z"/>
+</vector>
