+++ app/src/main/res/drawable-xhdpi/ic_search.xml
@@ -1,0 +1,12 @@
+<vector android:height="24dp" android:viewportHeight="512"
+    android:viewportWidth="512" android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#ff9e5e" android:pathData="m499.067,499.067c-17.244,17.244 -45.202,17.244 -62.446,0l-144.032,-144.032c-17.244,-17.244 -17.244,-45.202 0,-62.446 17.244,-17.244 45.202,-17.244 62.446,0l144.032,144.032c17.244,17.244 17.244,45.202 0,62.446z"/>
+    <path android:fillColor="#fa8d46" android:pathData="m499.067,436.621 l-144.032,-144.032c-14.03,-14.03 -35.153,-16.643 -51.827,-7.842 3.819,2.016 7.406,4.629 10.619,7.842l144.033,144.032c17.244,17.244 17.244,45.202 0,62.446 -3.214,3.214 -6.8,5.827 -10.619,7.842 16.674,8.801 37.796,6.188 51.827,-7.842 17.243,-17.244 17.243,-45.202 -0.001,-62.446z"/>
+    <path android:fillColor="#ff7893" android:pathData="m302.329,397.732c-12.018,-12.017 -12.018,-31.502 0,-43.519l51.884,-51.884c12.017,-12.017 31.502,-12.017 43.519,0 12.017,12.017 12.017,31.501 0,43.519l-51.884,51.884c-12.018,12.017 -31.502,12.017 -43.519,0z"/>
+    <path android:fillColor="#59a1a5" android:pathData="M200.77,200.77m-200.77,0a200.77,200.77 0,1 1,401.54 0a200.77,200.77 0,1 1,-401.54 0"/>
+    <path android:fillColor="#b5e8e0" android:pathData="M200.77,200.77m-149.261,0a149.261,149.261 0,1 1,298.522 0a149.261,149.261 0,1 1,-298.522 0"/>
+    <path android:fillColor="#97dbd1" android:pathData="m200.77,51.509c-6.99,0 -13.866,0.494 -20.604,1.429 72.586,10.067 128.657,72.52 128.657,147.832s-56.071,137.764 -128.657,147.832c6.737,0.934 13.613,1.429 20.604,1.429 82.302,0 149.261,-66.958 149.261,-149.261s-66.959,-149.261 -149.261,-149.261z"/>
+    <path android:fillColor="#fff7e8" android:pathData="M130.61,200.77m-21.382,0a21.382,21.382 0,1 1,42.764 0a21.382,21.382 0,1 1,-42.764 0"/>
+    <path android:fillColor="#fff7e8" android:pathData="M200.77,200.77m-21.382,0a21.382,21.382 0,1 1,42.764 0a21.382,21.382 0,1 1,-42.764 0"/>
+    <path android:fillColor="#fff7e8" android:pathData="M270.929,200.77m-21.382,0a21.382,21.382 0,1 1,42.764 0a21.382,21.382 0,1 1,-42.764 0"/>
+</vector>
