--- app/src/main/res/layout/fragment_dash_search.xml
@@ -1,26 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/clear_focus_view"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_gravity="center"
-    android:focusableInTouchMode="true"
-    android:orientation="vertical"
-    android:padding="8dp">
-
-    <Spinner
-        android:id="@+id/search_spin"
-        style="@style/SpinnerTheme"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content" />
-
-    <androidx.appcompat.widget.SearchView
-        android:id="@+id/search_view"
-        android:layout_width="match_parent"
-        android:layout_height="60dp"
-        android:layout_marginTop="10dp"
-        android:background="@color/app_bg_light2"
-        android:iconifiedByDefault="false"
-        android:queryHint="@string/search_hint" />
-</LinearLayout>
